requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
nltk>=3.6.0
openai>=1.0.0
